photo changes for new members

+++ src/app/admin/wallet-admin/wallet-admin.component.ts
@@ -105,21 +105,22 @@ export class WalletAdminComponent implements OnInit {
   }
 
   removeUser(userId: string, rowIndex: number): void {
-    confirm(
+    const res = confirm(
       this.translate.currentLang === 'en'
         ? 'Do you really want to remove this user from your wallet?'
         : 'Вы действительно хотите убрать этого пользователя из Вашего кошелька?'
     );
-    this.admService.removeUser(userId).subscribe(
-      (response) => {
-        this.users.data.splice(rowIndex, 1);
-        this.alertify.success(response);
-        this.users.data = this.users.data;
-      },
-      (error) => {
-        this.alertify.error(error.error);
-      }
-    );
+    if (res)
+      this.admService.removeUser(userId).subscribe(
+        (response) => {
+          this.users.data.splice(rowIndex, 1);
+          this.alertify.success(response);
+          this.users.data = this.users.data;
+        },
+        (error) => {
+          this.alertify.error(error.error);
+        }
+      );
   }
 
   sendInvitation(): void {
